+++ tests/jwk_test.py
@@ -22,6 +22,11 @@ RSA512_KEY = test_util.load_rsa_private_key('rsa512_key.pem')
 EC_P256_KEY = test_util.load_ec_private_key('ec_p256_key.pem')
 EC_P384_KEY = test_util.load_ec_private_key('ec_p384_key.pem')
 EC_P521_KEY = test_util.load_ec_private_key('ec_p521_key.pem')
+Ed25519_KEY = test_util.load_okp_private_key('ed25519_key.pem')
+Ed448_KEY = test_util.load_okp_private_key('ed448_key.pem')
+# Not implemented on my machine locally, and just cause barf from OpenSSL
+# X25519_KEY = test_util.load_okp_private_key('x25519_key.pem')
+# X448_KEY = test_util.load_okp_private_key('x448_key.pem')
 
 
 class JWKTest(unittest.TestCase):
@@ -336,6 +341,147 @@ AwEHoUQDQgAEGS5RvStca15z2FEanCM3juoX7tE/LB7iD44GWawGE40APAl/iZuH
         assert y[0] == 0
         assert len(y) == 32
         JWK.from_json(json)
+
+
+class JWKOKPTestBase(unittest.TestCase):
+    pass
+
+
+class JWKOKPTest(JWKOKPTestBase):
+    """Tests for josepy.jwk.JWKOKP."""
+
+    thumbprint = (
+        b'kPrK_qmxVWaYVA9wwBF6Iuo3vVzz7TxHCTwXBygrS4k'
+    )
+
+    def setUp(self):
+        from josepy.jwk import JWKOKP
+        self.ed25519_key = JWKOKP(key=Ed25519_KEY.public_key())
+        self.ed448_key = JWKOKP(key=Ed448_KEY.public_key())
+        # self.x25519_key = JWKOKP(key=X25519_KEY.public_key())
+        # self.x448_key = JWKOKP(key=X448_KEY.public_key())
+        # self.private = self.x448_key
+        # self.jwk = self.private
+        # Test vectors taken from RFC 8037, A.2
+        self.jwked25519json = {
+            'kty': 'OKP',
+            'crv': 'Ed25519',
+            'x': '11qYAYKxCrfVS_7TyWQHOg7hcvPapiMlrwIaaPcHURo',
+        }
+        self.jwked448json = {
+            'kty': 'OKP',
+            'crv': 'Ed448',
+            'x': (
+                "9b08f7cc31b7e3e67d22d5aea121074a273bd2b83de09c63faa73d2c"
+                "22c5d9bbc836647241d953d40c5b12da88120d53177f80e532c41fa0"
+            )
+        }
+        # Test vectors taken from
+        # https://datatracker.ietf.org/doc/html/rfc7748#section-6.1
+        self.jwkx25519json = {
+            'kty': 'OKP',
+            'crv': 'X25519',
+            'x': '8520f0098930a754748b7ddcb43ef75a0dbf3a0d26381af4eba4a98eaa9b4e6a',
+        }
+        # not 56 bytes long
+        self.jwkx448json = {
+            "kty": "OKP",
+            "crv": "X448",
+            "x": "jjQtV-fA7J_tK8dPzYq7jRPNjF8r5p6LW2R25S2Gw5U",
+        }
+        self.jwk = self.jwked25519json
+        self.private = JWKOKP(key=Ed25519_KEY)
+
+    def test_encode_ed448(self):
+        from josepy.jwk import JWKOKP
+        data = b"""-----BEGIN PRIVATE KEY-----
+MEcCAQAwBQYDK2VxBDsEOfqsAFWdop10FFPW7Ha2tx2AZh0Ii+jfL2wFXU/dY/fe
+iU7/vrGmQ+ux26NkgzfploOHZjEmltLJ9w==
+-----END PRIVATE KEY-----"""
+        key = JWKOKP.load(data)
+        partial = key.to_partial_json()
+        self.assertEqual(partial['crv'], 'Ed448')
+
+    def test_encode_ed25519(self):
+        from josepy.jwk import JWKOKP
+        data = b"""-----BEGIN PRIVATE KEY-----
+MC4CAQAwBQYDK2VwBCIEIPIAha9VqyHHpY1GtEW8JXWqLU5mrPRhXPwJqCtL3bWZ
+-----END PRIVATE KEY-----"""
+        key = JWKOKP.load(data)
+        data = key.to_partial_json()
+        self.assertEqual(data['x'], "9ujoz88QZL05w2lhaqUbBaBpwmM12Y7Y8Ybfwjibk-I")
+        self.assertEqual(data['d'], "8gCFr1WrIceljUa0RbwldaotTmas9GFc_AmoK0vdtZk")
+
+    def test_from_json(self):
+        from josepy.jwk import JWK
+        key = JWK.from_json(self.jwked25519json)
+        with self.subTest(key=[
+            self.jwked448json,
+            self.jwked25519json,
+        ]):
+            self.assertIsInstance(key.key, util.ComparableOKPKey)
+
+    def test_load(self):
+        from josepy.jwk import JWKOKP
+        self.assertEqual(
+            self.private,
+            JWKOKP.load(test_util.load_vector('ed25519_key.pem')),
+        )
+
+    def test_fields_to_json(self):
+        from josepy.jwk import JWK
+        data = b"""-----BEGIN PRIVATE KEY-----
+MC4CAQAwBQYDK2VwBCIEIPIAha9VqyHHpY1GtEW8JXWqLU5mrPRhXPwJqCtL3bWZ
+-----END PRIVATE KEY-----"""
+        key = JWK.load(data)
+        data = key.fields_to_partial_json()
+        self.assertEqual(data["crv"], "Ed25519")
+        self.assertEqual(data["d"], "8gCFr1WrIceljUa0RbwldaotTmas9GFc_AmoK0vdtZk")
+
+    @unittest.skip
+    def test_init_auto_comparable(self):
+        self.assertIsInstance(self.x448_key.key, util.ComparableOKPKey)
+
+    def test_unknown_crv_name(self):
+        from josepy.jwk import JWK
+        self.assertRaises(
+            errors.DeserializationError, JWK.from_json,
+            {
+                'kty': 'OKP',
+                'crv': 'Ed1000',
+                'x': 'jjQtV-fA7J_tK8dPzYq7jRPNjF8r5p6LW2R25S2Gw5U',
+            }
+        )
+
+    def test_no_x_value(self):
+        from josepy.jwk import JWK
+        with self.assertRaises(errors.DeserializationError) as warn:
+            JWK.from_json(
+                {
+                    "kty": "OKP",
+                    "crv": "Ed448",
+                }
+            )
+        self.assertEqual(
+            warn.exception.__str__(),
+            'Deserialization error: OKP should have "x" parameter'
+        )
+
+    def test_from_json_hashable(self):
+        from josepy.jwk import JWK
+        h = hash(JWK.from_json(self.jwked25519json))
+        self.assertIsInstance(h, int)
+
+    def test_deserialize_public_key(self):
+        # should target jwk.py:474-484, but those lines are still marked as missing
+        # in the coverage report
+        from josepy.jwk import JWKOKP
+        JWKOKP.fields_from_json(self.jwked25519json)
+
+    @unittest.skip
+    def test_x448(self):
+        from josepy.jwk import JWKOKP
+        _ = JWKOKP.fields_from_json(self.jwkx448json)
 
 
 if __name__ == '__main__':